Harden error handling in analysis.P2
Single trajectory analysis is susceptible to nan's in the MSD, which
can raise "invalid value in divide" and "divide by zero" errors. So far
these were ignored with the warnings module, because by default they are
warnings; however, even in code that uses "all='raise'" to make sure
that no errors go uncaught, we want to ignore these errors (since the
implicit behavior is correct). So now we also ignore these two issues
when raised as errors (only in P2 averaging)